requirements.txt
# ── Core ML ───────────────────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0

# ── Backend ───────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit>=1.32.0
requests>=2.31.0

# ── Training / utilities ──────────────────────────────────────────────────────
tqdm>=4.65.0
tensorboard>=2.14.0
pandas>=2.0.0
kaggle>=1.6.0
